s_tune update and flow rate submenu
In order to address some reported issues, it was necessary to restructure the s_tune menu and make it similar for all models of Crane.  Additionally, after determining the flow rate selection on s_tune made things a bit too 'busy' it has been moved to the new submenu x_tuneflow.

This should resolve the issue reported in #37 but will not yet be closed until verified.