Transferred Motor Testing to qc.g
Recreated qc.g for all crane models(just a different file header). config.g will no longer call any motor tests directly.  Only qc.g is called, and is currently commented out in this release.  In addition, the motor test section has been transferred to qc.g, prior to the movement of the toolhead to the middle of the bed.  This, however, is also commented out for now.  This resolves #48 and introduces the same basic qc config to all models.